Wall-E/requirements.txt
mysql-connector-python>=8.0.0
pandas>=1.3.0
numpy>=1.19.0
scikit-learn>=0.24.0
joblib>=1.0.0 